fix: リアルタイムデータをRealmではなくメモリで管理
