--- NearU/Model/EncountData.swift
@@ -1,16 +1,0 @@
-//
-//  EncountData.swift
-//  NearU
-//
-//  Created by  髙橋和 on 2024/10/10.
-//
-
-import Foundation
-import RealmSwift
-
-class EncountData: Object, Identifiable {
-    @Persisted(primaryKey: true) var userId: String = ""
-    @Persisted var date: Date = Date()
-    @Persisted var rssi: Int = 0
-}
-
